make image engine interface more consistent
support heterogeneous clusters

fix lint
