+++ cmd/sealer/cmd/cluster/run.go
@@ -26,8 +26,8 @@ import (
 	"github.com/sealerio/sealer/pkg/application"
 	clusterruntime "github.com/sealerio/sealer/pkg/cluster-runtime"
 	"github.com/sealerio/sealer/pkg/clusterfile"
-	v12 "github.com/sealerio/sealer/pkg/define/image/v1"
-	imagecommon "github.com/sealerio/sealer/pkg/define/options"
+	imagev1 "github.com/sealerio/sealer/pkg/define/image/v1"
+	"github.com/sealerio/sealer/pkg/define/options"
 	"github.com/sealerio/sealer/pkg/imagedistributor"
 	"github.com/sealerio/sealer/pkg/imageengine"
 	"github.com/sealerio/sealer/pkg/infradriver"
@@ -89,30 +89,31 @@ func NewRunCmd() *cobra.Command {
 				return runWithClusterfile(clusterFile, runFlags)
 			}
 
-			imageEngine, err := imageengine.NewImageEngine(imagecommon.EngineGlobalConfigurations{})
+			imageEngine, err := imageengine.NewImageEngine(options.EngineGlobalConfigurations{})
 			if err != nil {
 				return err
 			}
 
-			if err = imageEngine.Pull(&imagecommon.PullOptions{
+			id, err := imageEngine.Pull(&options.PullOptions{
 				Quiet:      false,
 				PullPolicy: "missing",
 				Image:      args[0],
 				Platform:   "local",
-			}); err != nil {
+			})
+			if err != nil {
 				return err
 			}
 
-			extension, err := imageEngine.GetSealerImageExtension(&imagecommon.GetImageAnnoOptions{ImageNameOrID: args[0]})
+			imageSpec, err := imageEngine.Inspect(&options.InspectOptions{ImageNameOrID: id})
 			if err != nil {
 				return fmt.Errorf("failed to get cluster image extension: %s", err)
 			}
 
-			if extension.Type == v12.AppInstaller {
+			if imageSpec.ImageExtension.Type == imagev1.AppInstaller {
 				app := v2.ConstructApplication(nil, runFlags.Cmds, runFlags.AppNames)
 
 				return installApplication(args[0], runFlags.CustomEnv,
-					app, extension, nil, imageEngine, runFlags.Mode)
+					app, imageSpec.ImageExtension, nil, imageEngine, runFlags.Mode)
 			}
 
 			clusterFromFlag, err := utils.ConstructClusterForRun(args[0], runFlags)
@@ -130,7 +131,7 @@ func NewRunCmd() *cobra.Command {
 				return err
 			}
 
-			return createNewCluster(imageEngine, cf, runFlags.Mode)
+			return createNewCluster(imageEngine, cf, imageSpec, runFlags.Mode)
 		},
 	}
 	runFlags = &types.RunFlags{}
@@ -187,36 +188,37 @@ func runWithClusterfile(clusterFile string, runFlags *types.RunFlags) error {
 
 	cf.SetCluster(*cluster)
 	imageName := cluster.Spec.Image
-	imageEngine, err := imageengine.NewImageEngine(imagecommon.EngineGlobalConfigurations{})
+	imageEngine, err := imageengine.NewImageEngine(options.EngineGlobalConfigurations{})
 	if err != nil {
 		return err
 	}
 
-	if err = imageEngine.Pull(&imagecommon.PullOptions{
+	id, err := imageEngine.Pull(&options.PullOptions{
 		Quiet:      false,
 		PullPolicy: "missing",
 		Image:      imageName,
 		Platform:   "local",
-	}); err != nil {
+	})
+	if err != nil {
 		return err
 	}
 
-	extension, err := imageEngine.GetSealerImageExtension(&imagecommon.GetImageAnnoOptions{ImageNameOrID: imageName})
+	imageSpec, err := imageEngine.Inspect(&options.InspectOptions{ImageNameOrID: id})
 	if err != nil {
 		return fmt.Errorf("failed to get cluster image extension: %s", err)
 	}
 
-	if extension.Type == v12.AppInstaller {
+	if imageSpec.ImageExtension.Type == imagev1.AppInstaller {
 		app := v2.ConstructApplication(cf.GetApplication(), cluster.Spec.CMD, cluster.Spec.APPNames)
 
 		return installApplication(imageName, runFlags.CustomEnv, app,
-			extension, cf.GetConfigs(), imageEngine, runFlags.Mode)
+			imageSpec.ImageExtension, cf.GetConfigs(), imageEngine, runFlags.Mode)
 	}
 
-	return createNewCluster(imageEngine, cf, runFlags.Mode)
+	return createNewCluster(imageEngine, cf, imageSpec, runFlags.Mode)
 }
 
-func createNewCluster(imageEngine imageengine.Interface, cf clusterfile.Interface, mode string) error {
+func createNewCluster(imageEngine imageengine.Interface, cf clusterfile.Interface, imageSpec *imagev1.ImageSpec, mode string) error {
 	cluster := cf.GetCluster()
 	infraDriver, err := infradriver.NewInfraDriver(&cluster)
 	if err != nil {
@@ -270,7 +272,7 @@ func createNewCluster(imageEngine imageengine.Interface, cf clusterfile.Interfac
 
 	runtimeConfig := &clusterruntime.RuntimeConfig{
 		Distributor:            distributor,
-		ImageEngine:            imageEngine,
+		ImageSpec:              imageSpec,
 		Plugins:                plugins,
 		ContainerRuntimeConfig: cluster.Spec.ContainerRuntime,
 	}
@@ -350,7 +352,7 @@ func loadToRegistry(infraDriver infradriver.InfraDriver, distributor imagedistri
 	return nil
 }
 
-func installApplication(appImageName string, envs []string, app *v2.Application, extension v12.ImageExtension, configs []v1.Config, imageEngine imageengine.Interface, mode string) error {
+func installApplication(appImageName string, envs []string, app *v2.Application, extension imagev1.ImageExtension, configs []v1.Config, imageEngine imageengine.Interface, mode string) error {
 	logrus.Infof("start to install application: %s", appImageName)
 
 	v2App, err := application.NewV2Application(app, extension)
