+++ pkg/define/image/v1/sealer_image_test.go
